+++ egov/egov-ptis/src/main/java/org/egov/ptis/domain/dao/property/PropertyTypeMasterDAO.java
@@ -57,7 +57,7 @@ public interface PropertyTypeMasterDAO {
 
 	public PropertyTypeMaster getPropertyTypeMasterByCode(String code);
 
-	public PropertyTypeMaster findById(Integer id, boolean lock);
+	public PropertyTypeMaster findById(Long id, boolean lock);
 
 	public List<PropertyTypeMaster> findAll();
